web/users/api.py
from rest_framework import permissions, viewsets, mixins, status
from .models import User

from .serializers import UserSerializer


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
